元件拆分整理
