+++ src/Header.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Header: React.FC<any> = (): JSX.Element => (
+  <div className="header">
+    <div className="close-icon" />
+  </div>
+);
+
+export default Header;
